+++ docs/assets/index.850b82d3.css
@@ -1,0 +1,1 @@
+._tachi-navbar-container_1fjw1_1{background-color:var(--default-bg-color);padding:.3rem 1rem;display:flex;align-items:center;flex-direction:row-reverse}._search-icon_1fjw1_9{display:flex}._light-theme-icon_1fjw1_13{display:flex;margin-right:4px}._dark-theme-icon_1fjw1_18{display:flex;margin-left:4px}._tachi-avatar-container_ekz7x_1{overflow:hidden;display:inline-block;cursor:pointer}:root{caret-color:transparent;--inline-height: 2rem;--inline-radius: .5rem;--capsule-raius: 2rem;--inline-wrapper-border-width: .1rem;--inline-focus-border-width: .15rem;--default-vertical-padding: .3rem;--default-pure-white: #ffffff;--default-pure-dark: #000000;--default-light-dark: #1a1a1a}:root [data-theme=light]{--default-font-color: #213547;--default-bg-color: #f1f1f1;--default-border-color: #d0d7de;--hover-border-color: #b2bec9}:root [data-theme=dark]{--default-font-color: #ffffff;--default-bg-color: #161b22;--default-border-color: #d0d7de;--hover-border-color: #eef1f3}._tachi-input-wrapper_1pg1f_28{height:var(--inline-height);border:var(--inline-wrapper-border-width) solid var(--default-border-color);border-radius:var(--inline-radius);overflow:hidden;box-sizing:content-box;display:flex;background-color:var(--default-bg-color);color:var(--default-font-color)}._tachi-input-focus-wrapper_1pg1f_39{border:var(--inline-focus-border-width) solid #0969da}._tachi-input-slot-container_1pg1f_43{display:flex;align-items:center}._tachi-input-prefix-container_1pg1f_48,._tachi-input-suffix-container_1pg1f_52{padding:0rem var(--default-vertical-padding)}._tachi-input_1pg1f_28{font-size:1rem;padding:0rem var(--default-vertical-padding)}input{background:none;outline:none;caret-color:auto}:root{caret-color:transparent;--inline-height: 2rem;--inline-radius: .5rem;--capsule-raius: 2rem;--inline-wrapper-border-width: .1rem;--inline-focus-border-width: .15rem;--default-vertical-padding: .3rem;--default-pure-white: #ffffff;--default-pure-dark: #000000;--default-light-dark: #1a1a1a}:root [data-theme=light]{--default-font-color: #213547;--default-bg-color: #f1f1f1;--default-border-color: #d0d7de;--hover-border-color: #b2bec9}:root [data-theme=dark]{--default-font-color: #ffffff;--default-bg-color: #161b22;--default-border-color: #d0d7de;--hover-border-color: #eef1f3}._tachi-switch-wrapper_1pzbq_28{height:var(--inline-height);width:calc(var(--inline-height) + var(--inline-height));border:var(--inline-wrapper-border-width) solid var(--default-border-color);border-radius:var(--capsule-raius);background-color:var(--default-bg-color);display:flex;align-items:center;cursor:pointer}._tachi-switch-neg-wrapper_1pzbq_39:hover,._tachi-switch-act-wrapper_1pzbq_43:hover{border-color:var(--hover-border-color)}._tachi-switch-gain_1pzbq_47{height:calc(var(--inline-height) - var(--default-vertical-padding));width:calc(var(--inline-height) - var(--default-vertical-padding));border-radius:var(--capsule-raius);display:flex;justify-content:center;align-items:center;color:var(--default-font-color)}._tachi-switch-gain-negative_1pzbq_57{background-color:var(--default-pure-white);animation:.1s _ani-gain-negative_1pzbq_1 forwards}._tachi-switch-gain-active_1pzbq_62{background-color:var(--default-pure-dark);animation:.1s _ani-gain-to-active_1pzbq_1 forwards}@keyframes _ani-gain-to-active_1pzbq_1{0%{transform:0}50%{transform:translate(var(--inline-height/2))}to{transform:translate(var(--inline-height))}}@keyframes _ani-gain-negative_1pzbq_1{0%{transform:translate(var(--inline-height))}50%{transform:translate(var(--inline-height/2))}to{transform:0}}*,:before,:after{box-sizing:border-box;border-width:0;border-style:solid;border-color:#e5e7eb}:before,:after{--tw-content: ""}html{line-height:1.5;-webkit-text-size-adjust:100%;-moz-tab-size:4;-o-tab-size:4;tab-size:4;font-family:ui-sans-serif,system-ui,-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Helvetica Neue,Arial,Noto Sans,sans-serif,"Apple Color Emoji","Segoe UI Emoji",Segoe UI Symbol,"Noto Color Emoji";font-feature-settings:normal}body{margin:0;line-height:inherit}hr{height:0;color:inherit;border-top-width:1px}abbr:where([title]){-webkit-text-decoration:underline dotted;text-decoration:underline dotted}h1,h2,h3,h4,h5,h6{font-size:inherit;font-weight:inherit}a{color:inherit;text-decoration:inherit}b,strong{font-weight:bolder}code,kbd,samp,pre{font-family:ui-monospace,SFMono-Regular,Menlo,Monaco,Consolas,Liberation Mono,Courier New,monospace;font-size:1em}small{font-size:80%}sub,sup{font-size:75%;line-height:0;position:relative;vertical-align:baseline}sub{bottom:-.25em}sup{top:-.5em}table{text-indent:0;border-color:inherit;border-collapse:collapse}button,input,optgroup,select,textarea{font-family:inherit;font-size:100%;font-weight:inherit;line-height:inherit;color:inherit;margin:0;padding:0}button,select{text-transform:none}button,[type=button],[type=reset],[type=submit]{-webkit-appearance:button;background-color:transparent;background-image:none}:-moz-focusring{outline:auto}:-moz-ui-invalid{box-shadow:none}progress{vertical-align:baseline}::-webkit-inner-spin-button,::-webkit-outer-spin-button{height:auto}[type=search]{-webkit-appearance:textfield;outline-offset:-2px}::-webkit-search-decoration{-webkit-appearance:none}::-webkit-file-upload-button{-webkit-appearance:button;font:inherit}summary{display:list-item}blockquote,dl,dd,h1,h2,h3,h4,h5,h6,hr,figure,p,pre{margin:0}fieldset{margin:0;padding:0}legend{padding:0}ol,ul,menu{list-style:none;margin:0;padding:0}textarea{resize:vertical}input::-moz-placeholder,textarea::-moz-placeholder{opacity:1;color:#9ca3af}input::placeholder,textarea::placeholder{opacity:1;color:#9ca3af}button,[role=button]{cursor:pointer}:disabled{cursor:default}img,svg,video,canvas,audio,iframe,embed,object{display:block;vertical-align:middle}img,video{max-width:100%;height:auto}[hidden]{display:none}*,:before,:after{--tw-border-spacing-x: 0;--tw-border-spacing-y: 0;--tw-translate-x: 0;--tw-translate-y: 0;--tw-rotate: 0;--tw-skew-x: 0;--tw-skew-y: 0;--tw-scale-x: 1;--tw-scale-y: 1;--tw-pan-x: ;--tw-pan-y: ;--tw-pinch-zoom: ;--tw-scroll-snap-strictness: proximity;--tw-ordinal: ;--tw-slashed-zero: ;--tw-numeric-figure: ;--tw-numeric-spacing: ;--tw-numeric-fraction: ;--tw-ring-inset: ;--tw-ring-offset-width: 0px;--tw-ring-offset-color: #fff;--tw-ring-color: rgb(59 130 246 / .5);--tw-ring-offset-shadow: 0 0 #0000;--tw-ring-shadow: 0 0 #0000;--tw-shadow: 0 0 #0000;--tw-shadow-colored: 0 0 #0000;--tw-blur: ;--tw-brightness: ;--tw-contrast: ;--tw-grayscale: ;--tw-hue-rotate: ;--tw-invert: ;--tw-saturate: ;--tw-sepia: ;--tw-drop-shadow: ;--tw-backdrop-blur: ;--tw-backdrop-brightness: ;--tw-backdrop-contrast: ;--tw-backdrop-grayscale: ;--tw-backdrop-hue-rotate: ;--tw-backdrop-invert: ;--tw-backdrop-opacity: ;--tw-backdrop-saturate: ;--tw-backdrop-sepia: }::backdrop{--tw-border-spacing-x: 0;--tw-border-spacing-y: 0;--tw-translate-x: 0;--tw-translate-y: 0;--tw-rotate: 0;--tw-skew-x: 0;--tw-skew-y: 0;--tw-scale-x: 1;--tw-scale-y: 1;--tw-pan-x: ;--tw-pan-y: ;--tw-pinch-zoom: ;--tw-scroll-snap-strictness: proximity;--tw-ordinal: ;--tw-slashed-zero: ;--tw-numeric-figure: ;--tw-numeric-spacing: ;--tw-numeric-fraction: ;--tw-ring-inset: ;--tw-ring-offset-width: 0px;--tw-ring-offset-color: #fff;--tw-ring-color: rgb(59 130 246 / .5);--tw-ring-offset-shadow: 0 0 #0000;--tw-ring-shadow: 0 0 #0000;--tw-shadow: 0 0 #0000;--tw-shadow-colored: 0 0 #0000;--tw-blur: ;--tw-brightness: ;--tw-contrast: ;--tw-grayscale: ;--tw-hue-rotate: ;--tw-invert: ;--tw-saturate: ;--tw-sepia: ;--tw-drop-shadow: ;--tw-backdrop-blur: ;--tw-backdrop-brightness: ;--tw-backdrop-contrast: ;--tw-backdrop-grayscale: ;--tw-backdrop-hue-rotate: ;--tw-backdrop-invert: ;--tw-backdrop-opacity: ;--tw-backdrop-saturate: ;--tw-backdrop-sepia: }.static{position:static}.rounded-full{border-radius:9999px}
