+++ docs/index.html
@@ -5,11 +5,13 @@
     <link rel="icon" href="/assets/sakurajima.44cdfb32.jpeg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Tachish's Blog</title>
-    <script type="module" crossorigin src="/assets/index.3a06d488.js"></script>
-    <link rel="stylesheet" href="/assets/index.fac74221.css">
+    <script type="module" crossorigin src="/assets/index.db99855c.js"></script>
+    <link rel="stylesheet" href="/assets/index.850b82d3.css">
   </head>
   <body>
     <div id="app"></div>
+    <link rel='stylesheet' href='https://cdn-uicons.flaticon.com/uicons-regular-straight/css/uicons-regular-straight.css'>
+    <link rel='stylesheet' href='https://cdn-uicons.flaticon.com/uicons-regular-rounded/css/uicons-regular-rounded.css'>
     
   </body>
 </html>
