+++ package.json
@@ -6,6 +6,7 @@
   "scripts": {
     "dev": "vite",
     "fix": "eslint './src/**' --fix",
+    "simple-build": "vite build",
     "build": "vue-tsc --noEmit --skipLibCheck && vite build",
     "preview": "vite preview"
   },
